fixed unintentional casting between double and float

+++ headers/ray.hpp
@@ -12,7 +12,7 @@ public:
     const point3& origin() const;
     const vec3& direction() const;
 
-    point3 at(double t) const;
+    point3 at(float t) const;
 
 private:
     point3 orig;
